requirements-dev.txt
appdirs==1.4.3
automium==0.2.6
automium-web==0.1.1
boto3==1.4.4
botocore==1.5.78
click==6.7
dj-database-url==0.4.2
dj-static==0.0.6
Django==1.11.1
django-cleanup==1.0.0
django-debug-toolbar==1.8
django-recaptcha==1.3.1
django-storages==1.6.3
django-uuslug==1.1.8
djangorestframework==3.6.3
docutils==0.13.1
Flask==0.12.2
itsdangerous==0.24
Jinja2==2.9.6
jmespath==0.9.3
legacy==0.1.6
MarkupSafe==1.0
olefile==0.44
packaging==16.8
Pillow==4.2.0
psycopg2==2.7.1
pyparsing==2.2.0
python-dateutil==2.6.0
python-decouple==3.0
python-slugify==1.2.4
pytz==2017.2
s3transfer==0.1.10
six==1.10.0
sqlparse==0.2.3
static3==0.7.0
Unidecode==0.4.20
Werkzeug==0.12.2
